requirements.txt
Scrapy==1.6.0
setuptools==40.8.0
w3lib==1.20.0
orator==0.9.8
PyYAML==5.1